requirements.txt
fastapi
sqlalchemy
pydantic
databases
passlib
python-jose